requirements.txt
setuptools==46.2.0
wheel==0.34.2

google-cloud-secret-manager==2.6.0
google-cloud-storage==1.41.1
psycopg2-binary==2.9.1
python-dateutil==2.8.1
pyyaml==5.3.1
pandas==1.3.4
numpy==1.21.4
openpyxl==3.0.9
chardet==4.0.0
xlrd